Add function: copy/merge data from created excels.
Add a function that copy and merge data from range of created excels
specified by user. After user specifies range with the format of
range address of excel, the program will automatically copy data of
created excels by 'text_to_excel' function in the specified range
and merge them into a new excel file. User can determine the name
of the file. The new excel file will be created in the same
directory with text files containing raw data.

1. basic.file.__init__.py: merge_specified_range
This function is the key function of this update. It opens the
created excels and reads the data in the range specified by user.
The read data will be copied and merged into a new excel book.
The new excel book will be saved with the name and the path
determined by user.

2. gui.copiedrange
This module has a widget that can get an input of the range from
user.

3. gui.dialogs.DlgSaveFileName: __bt_start_load_clicked
Get an input for the name of the new excel file containing copied
data and excute 'merge_specified_range'.

4. gui.mainwindow: MainWindow
Add a widget from "gui.copiedrange" module.