Fix - erasing template / copying formula
There was a bug that the program saves excels created from text
files with deletion of data in the template excel file. There was
also a bug that the program copies formula (not values) when
merging data from the excels created from the text files. In this
commit, those two bugs are fixed.